add checklogut function on all final page
